좋아요 리스트 api 개발

+++ src/main/java/com/sktechx/godmusic/personal/rest/model/vo/preference/Artist.java
@@ -55,6 +55,6 @@ public class Artist {
         @ApiModelProperty(value = "이미지URL")
         private String url;
         @ApiModelProperty(value = "이미지사이즈")
-        private Integer size;
+        private Long size;
     }
 }
